Expect messages when creating new teast.FakeModel
It is difficult to use a well structured table-testing approach when the
messages a fake model receives are declared separately to its creation.
This also allows for tests to miss potentially expected behaviour.

A new teast.FakeModelOption called ExpectMessages allows us to specify
which messages a declared teast.FakeModel should received at the point
the fake is created.