Refactor stack test to use correct testing.T
Fake models were being created with the `t *testing.T` value from the
main test. This meant that when they failed in the subtest it was not
possible to indicate which subtest was failing based on a failed
expected message assertion.

It's not elegant, but providing a function to build the params allows us
to provide the appropriate `*testing.T` object.